Added MOSTheme provider, moved types
