+++ package.json
@@ -59,8 +59,7 @@
     "@types/react": "^16.3.5",
     "@types/react-dom": "^16.0.4",
     "react": "^16.3.1",
-    "react-dom": "^16.3.1",
-    "styled-components": "^3.2.4"
+    "react-dom": "^16.3.1"
   },
   "scripts": {
     "test": "jest --config=./config/jest/jest.config.json --coverage",
