+++ src/index.ts
@@ -6,3 +6,4 @@ export { default as Icon } from './components/Icon';
 export { default as Spacer } from './components/Spacer';
 export { default as SubHeading } from './components/SubHeading';
 export { default as Text } from './components/Text';
+export { default as MosTheme } from './themes/MosTheme';
