+++ src/themes/MosTheme/index.ts
@@ -1,0 +1,1 @@
+export { default as default } from './MosTheme';
