+++ src/types/ButtonThemeType.ts
@@ -1,0 +1,24 @@
+type ButtonThemeType = {
+    backgroundColor:{
+        default:string;
+        hover:string;
+        focus:string;
+        active:string;
+    };
+    borderColor:{
+        default:string;
+        hover:string;
+        focus:string;
+        active:string;
+    };
+    borderRadius:string;
+    borderWidth:string;
+    color:{
+        default:string;
+        hover:string;
+        focus:string;
+        active:string;
+    };
+};
+
+export default ButtonThemeType;
